[Optimization] Change size to more adequate storage type

+++ _install/data/sql/MySQL/pH7_Core.sql
@@ -1160,7 +1160,7 @@ INSERT INTO ph7_custom_code VALUES
 
 
 CREATE TABLE IF NOT EXISTS ph7_block_countries (
-  countryId smallint(4) unsigned NOT NULL AUTO_INCREMENT,
+  countryId tinyint(3) unsigned NOT NULL AUTO_INCREMENT,
   countryCode char(2) NOT NULL,
   PRIMARY KEY (countryId),
   UNIQUE KEY (countryCode)
